Fix up the Windows build

+++ src/struct.c
@@ -39,14 +39,14 @@
 
 #if (LUA_VERSION_NUM >= 502)
 
-#define luaL_register(L,n,f)	luaL_newlib(L,f)
+#define luaL_register(L,n,f) luaL_newlib(L,f)
 
 #endif
 
 
 /* basic integer type */
 #if !defined(STRUCT_INT)
-#define STRUCT_INT	long long
+#define STRUCT_INT long long
 #endif
 
 typedef STRUCT_INT Inttype;
@@ -56,10 +56,10 @@ typedef unsigned STRUCT_INT Uinttype;
 
 
 /* maximum size (in bytes) for integral types */
-#define MAXINTSIZE	32
+#define MAXINTSIZE 32
 
 /* is 'x' a power of 2? */
-#define isp2(x)		((x) > 0 && ((x) & ((x) - 1)) == 0)
+#define isp2(x) ((x) > 0 && ((x) & ((x) - 1)) == 0)
 
 /* dummy structure to get alignment requirements */
 struct cD {
@@ -68,19 +68,19 @@ struct cD {
 };
 
 
-#define PADDING		(sizeof(struct cD) - sizeof(double))
-#define MAXALIGN  	(PADDING > sizeof(int) ? PADDING : sizeof(int))
+#define PADDING   (sizeof(struct cD) - sizeof(double))
+#define MAXALIGN  (PADDING > sizeof(int) ? PADDING : sizeof(int))
 
 
 /* endian options */
-#define BIG	0
-#define LITTLE	1
+#define BIG     0
+#define LITTLE  1
 
 
 static union {
   int dummy;
   char endian;
-} const native = {1};
+} const native = { 1 };
 
 
 typedef struct Header {
@@ -89,41 +89,43 @@ typedef struct Header {
 } Header;
 
 
-static int getnum (const char **fmt, int df) {
+static int getnum(const char** fmt, int df)
+{
   if (!isdigit(**fmt))  /* no number? */
     return df;  /* return default value */
   else {
     int a = 0;
     do {
-      a = a*10 + *((*fmt)++) - '0';
-    } while (isdigit(**fmt));
+      a = a * 10 + *((*fmt)++) - '0';
+    }
+    while (isdigit(**fmt));
     return a;
   }
 }
 
 
-#define defaultoptions(h)	((h)->endian = native.endian, (h)->align = 1)
+#define defaultoptions(h) ((h)->endian = native.endian, (h)->align = 1)
 
 
 
-static size_t optsize (lua_State *L, char opt, const char **fmt) {
+static size_t optsize(lua_State* L, char opt, const char** fmt)
+{
   switch (opt) {
-    case 'B': case 'b': return sizeof(char);
-    case 'H': case 'h': return sizeof(short);
-    case 'L': case 'l': return sizeof(long);
-    case 'T': return sizeof(size_t);
-    case 'f':  return sizeof(float);
-    case 'd':  return sizeof(double);
-    case 'x': return 1;
-    case 'c': return getnum(fmt, 1);
-    case 'i': case 'I': {
+  case 'B': case 'b': return sizeof(char);
+  case 'H': case 'h': return sizeof(short);
+  case 'L': case 'l': return sizeof(long);
+  case 'T': return sizeof(size_t);
+  case 'f':  return sizeof(float);
+  case 'd':  return sizeof(double);
+  case 'x': return 1;
+  case 'c': return getnum(fmt, 1);
+  case 'i': case 'I': {
       int sz = getnum(fmt, sizeof(int));
-      if (sz > MAXINTSIZE)
-        luaL_error(L, "integral size %d is larger than limit of %d",
-                       sz, MAXINTSIZE);
+      if (sz > MAXINTSIZE) luaL_error(L, "integral size %d is larger than limit of %d",
+                                      sz, MAXINTSIZE);
       return sz;
     }
-    default: return 0;  /* other cases do not need alignment */
+  default: return 0;  /* other cases do not need alignment */
   }
 }
 
@@ -132,55 +134,53 @@ static size_t optsize (lua_State *L, char opt, const char **fmt) {
 ** return number of bytes needed to align an element of size 'size'
 ** at current position 'len'
 */
-static int gettoalign (size_t len, Header *h, int opt, size_t size) {
+static int gettoalign(size_t len, Header* h, int opt, size_t size)
+{
   if (size == 0 || opt == 'c') return 0;
-  if (size > (size_t)h->align)
-    size = h->align;  /* respect max. alignment */
-  return (size - (len & (size - 1))) & (size - 1);
+  if (size > (size_t)h->align) size = h->align;  /* respect max. alignment */
+  return (int)(size - (len & (size - 1))) & (size - 1);
 }
 
 
 /*
 ** options to control endianess and alignment
 */
-static void controloptions (lua_State *L, int opt, const char **fmt,
-                            Header *h) {
+static void controloptions(lua_State* L, int opt, const char** fmt,
+                           Header* h)
+{
   switch (opt) {
-    case  ' ': return;  /* ignore white spaces */
-    case '>': h->endian = BIG; return;
-    case '<': h->endian = LITTLE; return;
-    case '!': {
+  case  ' ': return;  /* ignore white spaces */
+  case '>': h->endian = BIG; return;
+  case '<': h->endian = LITTLE; return;
+  case '!': {
       int a = getnum(fmt, MAXALIGN);
-      if (!isp2(a))
-        luaL_error(L, "alignment %d is not a power of 2", a);
+      if (!isp2(a)) luaL_error(L, "alignment %d is not a power of 2", a);
       h->align = a;
       return;
     }
-    default: {
-      const char *msg = lua_pushfstring(L, "invalid format option '%c'", opt);
+  default: {
+      const char* msg = lua_pushfstring(L, "invalid format option '%c'", opt);
       luaL_argerror(L, 1, msg);
     }
   }
 }
 
 
-static void putinteger (lua_State *L, luaL_Buffer *b, int arg, int endian,
-                        int size) {
+static void putinteger(lua_State* L, luaL_Buffer* b, int arg, int endian,
+                       int size)
+{
   lua_Number n = luaL_checknumber(L, arg);
   Uinttype value;
   char buff[MAXINTSIZE];
-  if (n < 0)
-    value = (Uinttype)(Inttype)n;
-  else
-    value = (Uinttype)n;
+  if (n < 0) value = (Uinttype)(Inttype)n;
+  else value = (Uinttype)n;
   if (endian == LITTLE) {
     int i;
     for (i = 0; i < size; i++) {
       buff[i] = (value & 0xff);
       value >>= 8;
     }
-  }
-  else {
+  } else {
     int i;
     for (i = size - 1; i >= 0; i--) {
       buff[i] = (value & 0xff);
@@ -191,7 +191,8 @@ static void putinteger (lua_State *L, luaL_Buffer *b, int arg, int endian,
 }
 
 
-static void correctbytes (char *b, int size, int endian) {
+static void correctbytes(char* b, int size, int endian)
+{
   if (endian != native.endian) {
     int i = 0;
     while (i < --size) {
@@ -203,9 +204,10 @@ static void correctbytes (char *b, int size, int endian) {
 }
 
 
-static int b_pack (lua_State *L) {
+static int b_pack(lua_State* L)
+{
   luaL_Buffer b;
-  const char *fmt = luaL_checkstring(L, 1);
+  const char* fmt = luaL_checkstring(L, 1);
   Header h;
   int arg = 2;
   size_t totalsize = 0;
@@ -219,30 +221,30 @@ static int b_pack (lua_State *L) {
     totalsize += toalign;
     while (toalign-- > 0) luaL_addchar(&b, '\0');
     switch (opt) {
-      case 'b': case 'B': case 'h': case 'H':
-      case 'l': case 'L': case 'T': case 'i': case 'I': {  /* integer types */
-        putinteger(L, &b, arg++, h.endian, size);
+    case 'b': case 'B': case 'h': case 'H':
+    case 'l': case 'L': case 'T': case 'i': case 'I': {  /* integer types */
+        putinteger(L, &b, arg++, h.endian, (int)size);
         break;
       }
-      case 'x': {
+    case 'x': {
         luaL_addchar(&b, '\0');
         break;
       }
-      case 'f': {
+    case 'f': {
         float f = (float)luaL_checknumber(L, arg++);
-        correctbytes((char *)&f, size, h.endian);
-        luaL_addlstring(&b, (char *)&f, size);
+        correctbytes((char*)&f, (int)size, h.endian);
+        luaL_addlstring(&b, (char*)&f, size);
         break;
       }
-      case 'd': {
+    case 'd': {
         double d = luaL_checknumber(L, arg++);
-        correctbytes((char *)&d, size, h.endian);
-        luaL_addlstring(&b, (char *)&d, size);
+        correctbytes((char*)&d, (int)size, h.endian);
+        luaL_addlstring(&b, (char*)&d, size);
         break;
       }
-      case 'c': case 's': {
+    case 'c': case 's': {
         size_t l;
-        const char *s = luaL_checklstring(L, arg++, &l);
+        const char* s = luaL_checklstring(L, arg++, &l);
         if (size == 0) size = l;
         luaL_argcheck(L, l >= (size_t)size, arg, "string too short");
         luaL_addlstring(&b, s, size);
@@ -252,7 +254,7 @@ static int b_pack (lua_State *L) {
         }
         break;
       }
-      default: controloptions(L, opt, &fmt, &h);
+    default: controloptions(L, opt, &fmt, &h);
     }
     totalsize += size;
   }
@@ -261,8 +263,9 @@ static int b_pack (lua_State *L) {
 }
 
 
-static lua_Number getinteger (const char *buff, int endian,
-                        int issigned, int size) {
+static lua_Number getinteger(const char* buff, int endian,
+                             int issigned, int size)
+{
   Uinttype l = 0;
   int i;
   if (endian == BIG) {
@@ -270,17 +273,15 @@ static lua_Number getinteger (const char *buff, int endian,
       l <<= 8;
       l |= (Uinttype)(unsigned char)buff[i];
     }
-  }
-  else {
+  } else {
     for (i = size - 1; i >= 0; i--) {
       l <<= 8;
       l |= (Uinttype)(unsigned char)buff[i];
     }
   }
-  if (!issigned)
-    return (lua_Number)l;
+  if (!issigned) return (lua_Number)l;
   else {  /* signed format */
-    Uinttype mask = (Uinttype)(~((Uinttype)0)) << (size*8 - 1);
+    Uinttype mask = (Uinttype)(~((Uinttype)0)) << (size * 8 - 1);
     if (l & mask)  /* negative value? */
       l |= mask;  /* signal extension */
     return (lua_Number)(Inttype)l;
@@ -288,11 +289,12 @@ static lua_Number getinteger (const char *buff, int endian,
 }
 
 
-static int b_unpack (lua_State *L) {
+static int b_unpack(lua_State* L)
+{
   Header h;
-  const char *fmt = luaL_checkstring(L, 1);
+  const char* fmt = luaL_checkstring(L, 1);
   size_t ld;
-  const char *data = luaL_checklstring(L, 2, &ld);
+  const char* data = luaL_checklstring(L, 2, &ld);
   size_t pos = luaL_optinteger(L, 3, 1) - 1;
   defaultoptions(&h);
   lua_settop(L, 2);
@@ -300,53 +302,51 @@ static int b_unpack (lua_State *L) {
     int opt = *fmt++;
     size_t size = optsize(L, opt, &fmt);
     pos += gettoalign(pos, &h, opt, size);
-    luaL_argcheck(L, pos+size <= ld, 2, "data string too short");
+    luaL_argcheck(L, pos + size <= ld, 2, "data string too short");
     luaL_checkstack(L, 1, "too many results");
     switch (opt) {
-      case 'b': case 'B': case 'h': case 'H':
-      case 'l': case 'L': case 'T': case 'i':  case 'I': {  /* integer types */
+    case 'b': case 'B': case 'h': case 'H':
+    case 'l': case 'L': case 'T': case 'i':  case 'I': {  /* integer types */
         int issigned = islower(opt);
-        lua_Number res = getinteger(data+pos, h.endian, issigned, size);
+        lua_Number res = getinteger(data + pos, h.endian, issigned, (int)size);
         lua_pushnumber(L, res);
         break;
       }
-      case 'x': {
+    case 'x': {
         break;
       }
-      case 'f': {
+    case 'f': {
         float f;
-        memcpy(&f, data+pos, size);
-        correctbytes((char *)&f, sizeof(f), h.endian);
+        memcpy(&f, data + pos, size);
+        correctbytes((char*)&f, sizeof(f), h.endian);
         lua_pushnumber(L, f);
         break;
       }
-      case 'd': {
+    case 'd': {
         double d;
-        memcpy(&d, data+pos, size);
-        correctbytes((char *)&d, sizeof(d), h.endian);
+        memcpy(&d, data + pos, size);
+        correctbytes((char*)&d, sizeof(d), h.endian);
         lua_pushnumber(L, d);
         break;
       }
-      case 'c': {
+    case 'c': {
         if (size == 0) {
-          if (!lua_isnumber(L, -1))
-            luaL_error(L, "format `c0' needs a previous size");
-          size = lua_tonumber(L, -1);
+          if (!lua_isnumber(L, -1)) luaL_error(L, "format `c0' needs a previous size");
+          size = (unsigned)lua_tonumber(L, -1);
           lua_pop(L, 1);
-          luaL_argcheck(L, pos+size <= ld, 2, "data string too short");
+          luaL_argcheck(L, pos + size <= ld, 2, "data string too short");
         }
-        lua_pushlstring(L, data+pos, size);
+        lua_pushlstring(L, data + pos, size);
         break;
       }
-      case 's': {
-        const char *e = (const char *)memchr(data+pos, '\0', ld - pos);
-        if (e == NULL)
-          luaL_error(L, "unfinished string in data");
-        size = (e - (data+pos)) + 1;
-        lua_pushlstring(L, data+pos, size - 1);
+    case 's': {
+        const char* e = (const char*)memchr(data + pos, '\0', ld - pos);
+        if (e == NULL) luaL_error(L, "unfinished string in data");
+        size = (e - (data + pos)) + 1;
+        lua_pushlstring(L, data + pos, size - 1);
         break;
       }
-      default: controloptions(L, opt, &fmt, &h);
+    default: controloptions(L, opt, &fmt, &h);
     }
     pos += size;
   }
@@ -355,21 +355,19 @@ static int b_unpack (lua_State *L) {
 }
 
 
-static int b_size (lua_State *L) {
+static int b_size(lua_State* L)
+{
   Header h;
-  const char *fmt = luaL_checkstring(L, 1);
+  const char* fmt = luaL_checkstring(L, 1);
   size_t pos = 0;
   defaultoptions(&h);
   while (*fmt) {
     int opt = *fmt++;
     size_t size = optsize(L, opt, &fmt);
     pos += gettoalign(pos, &h, opt, size);
-    if (opt == 's')
-      luaL_argerror(L, 1, "option 's' has no fixed size");
-    else if (opt == 'c' && size == 0)
-      luaL_argerror(L, 1, "option 'c0' has no fixed size");
-    if (!isalnum(opt))
-      controloptions(L, opt, &fmt, &h);
+    if (opt == 's') luaL_argerror(L, 1, "option 's' has no fixed size");
+    else if (opt == 'c' && size == 0) luaL_argerror(L, 1, "option 'c0' has no fixed size");
+    if (!isalnum(opt)) controloptions(L, opt, &fmt, &h);
     pos += size;
   }
   lua_pushinteger(L, pos);
@@ -381,16 +379,17 @@ static int b_size (lua_State *L) {
 
 
 static const struct luaL_Reg thislib[] = {
-  {"pack", b_pack},
-  {"unpack", b_unpack},
-  {"size", b_size},
-  {NULL, NULL}
+  { "pack", b_pack },
+  { "unpack", b_unpack },
+  { "size", b_size },
+  { NULL, NULL }
 };
 
 
-LUALIB_API int luaopen_struct (lua_State *L);
+LUALIB_API int luaopen_struct(lua_State* L);
 
-LUALIB_API int luaopen_struct (lua_State *L) {
+LUALIB_API int luaopen_struct(lua_State* L)
+{
   luaL_register(L, "struct", thislib);
   return 1;
 }
